pyproject.toml
[tool.poetry]
name = "galley"
version = "0.1.0"
description = ""
authors = ["David Francos <davidfrancos@sipay.es>"]

[tool.poetry.dependencies]
python = "^3.10"
aiohttp-swagger3 = "^0.7.2"
aiohttp = "^3.8.1"
cleo = "^0.8.1"
loguru = "^0.6.0"
tortoise-orm = "^0.19.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
galley = "galley:run_app"
